requirements.txt
requests==2.18.2
python-whois==0.6.5
